merge jobs `lint` and `tests`
These 2 sequential jobs were higly similar, the only difference being
the last step: the lint vs the test suite.

This a waste of resources as they will create a new runner.

Instead we run them in the same job/same runner, and because GitHub
Actions steps are sequential too, the step for the tests suite will only
run if the lint step is successful.