Stop using Ember.A()
Ember's Array prototype extensions are being deprecated soon, and
Ember Data has already deprecated them. As part of Ember Data's
deprecation work calling Ember.A on a PromiseMany has been deprecated so
removing any calls to this on passed values in composable helpers is
needed.

For the most part the usage of Ember.A here were replaced with native
Array.includes which is supported everywhere but IE 11

unfortunately Array.at() does not yet have universal support so I've
added a polyfill for that copied from the TC39 proposal.